parisc/unwind: fix unwinder when CONFIG_64BIT is enabled
Git-commit: a1ec31a0befa5458366959224e8e6d94c1513fa1
References: bsn#19
Patch-mainline: v5.10.80

[ Upstream commit 8e0ba125c2bf1030af3267058019ba86da96863f ]

With 64 bit kernels unwind_special() is not working because
it compares the pc to the address of the function descriptor.
Add a helper function that compares pc with the dereferenced
address. This fixes all of the backtraces on my c8000. Without
this changes, a lot of backtraces are missing in kdb or the
show-all-tasks command from /proc/sysrq-trigger.

Signed-off-by: Sven Schnelle <svens@stackframe.org>
Signed-off-by: Helge Deller <deller@gmx.de>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.80-0255-parisc-unwind-fix-unwinder-when-CONFIG_64BIT-is-enab.patch